[1.2.1fix many bugs
