+++ src/main/java/com/hippo/fresh/FreshApplication.java
@@ -39,10 +39,10 @@ public class FreshApplication {
 //	}
 //
 ////	从数据库中读出商品信息写到ES中
-//	@PostConstruct
-//	public void buildIndex() {
-////		elasticsearchOperations.indexOps(SearchProduct.class).refresh();
-////		searchProductRepository.deleteAll();
-//		searchProductRepository.saveAll(ScheduleConfig.ConvertProduct(productRepository.findAll()));
-//	}
+	@PostConstruct
+	public void buildIndex() {
+//		elasticsearchOperations.indexOps(SearchProduct.class).refresh();
+//		searchProductRepository.deleteAll();
+		searchProductRepository.saveAll(ScheduleConfig.ConvertProduct(productRepository.findAll()));
+	}
 }
